refactor: 이름 객체 삭제
