+++ src/main/java/ladder/domain/Player.java
@@ -2,25 +2,30 @@ package ladder.domain;
 
 public class Player {
 
+    private static final int MAX_NAME_LENGTH = 5;
+
     private int position;
-    private final Name name;
+    private final String name;
 
     public Player(final String name) {
-        this.name = new Name(name);
+        validateName(name);
+        this.name = name;
         this.position = 0;
     }
 
     public Player(final String name, final int position) {
-        this.name = new Name(name);
+        validateName(name);
+        this.name = name;
         this.position = position;
     }
 
-    public String getName() {
-        return name.getValue();
-    }
-
-    public int getPosition() {
-        return position;
+    private void validateName(final String name) {
+        if (name.length() > MAX_NAME_LENGTH) {
+            throw new IllegalArgumentException("참여자의 이름은 최대 " + MAX_NAME_LENGTH + "글자를 넘을 수 없습니다.\n" + "Name : " + name);
+        }
+        if (!name.replace(" ", "").equals(name)) {
+            throw new IllegalArgumentException("참여자의 이름에는 공백이 들어갈 수 없습니다.\n" + "Name : " + name);
+        }
     }
 
     public void move(Direction direction) {
@@ -32,4 +37,13 @@ public class Player {
             this.position++;
         }
     }
+
+    public String getName() {
+        return name;
+    }
+
+    public int getPosition() {
+        return position;
+    }
+
 }
