+++ src/main/java/ladder/domain/Product.java
@@ -1,14 +1,15 @@
 package ladder.domain;
 
+// TODO : Getter의 성질만 가지고있다는 부분에 대한 의문 가져보기
 public class Product {
 
-    private final Name name;
+    private final String name;
 
     public Product(final String name) {
-        this.name = new Name(name);
+        this.name = name;
     }
 
     public String getName() {
-        return name.getValue();
+        return name;
     }
 }
