+++ src/test/java/ladder/domain/PlayerTest.java
@@ -6,6 +6,7 @@ import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
 class PlayerTest {
@@ -14,6 +15,20 @@ class PlayerTest {
     void 참여자는_이름을_가질_수_있다() {
         Player player = new Player("이름", 0);
         assertThat(player.getName()).isEqualTo("이름");
+    }
+
+    @Test
+    void 참여자의_이름은_5자를_초과할_수_없다() {
+        assertThatThrownBy(() -> {
+            new Player("5글자넘어가는이름");
+        }).isInstanceOf(IllegalArgumentException.class);
+    }
+
+    @Test
+    void 참여자의_이름에는_공백이_들어갈_수_없다() {
+        assertThatThrownBy(() -> {
+            new Player("김 한국");
+        }).isInstanceOf(IllegalArgumentException.class);
     }
 
     @Nested
